# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: 新点 SaaS 造价系统 API
  version: 1.0.0-current
  summary: 当前代码已注册的 V1 API 最小契约
  description: |
    自动生成文件，请勿手工编辑。
    生成命令：npm run docs:openapi-current
servers:
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    ObjectEnvelope:
      type: object
      additionalProperties: true
    ListEnvelope:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
      additionalProperties: true
paths:
...
  /v1/projects/{projectId}/bill-versions/{billVersionId}/items/{itemId}/quota-lines:
    get:
      operationId: getProjectsByProjectIdBillVersionsByBillVersionIdItemsByItemIdQuotaLines
      summary: GET /v1/projects/:projectId/bill-versions/:billVersionId/items/:itemId/quota-lines
      tags:
        - Quota Lines
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      - name: billVersionId
        in: path
        description: Bill version id.
        required: true
        schema:
          type: string
      - name: itemId
        in: path
        description: itemId path parameter.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: postProjectsByProjectIdBillVersionsByBillVersionIdItemsByItemIdQuotaLines
      summary: POST /v1/projects/:projectId/bill-versions/:billVersionId/items/:itemId/quota-lines
      tags:
        - Quota Lines
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      - name: billVersionId
        in: path
        description: Bill version id.
        required: true
        schema:
          type: string
      - name: itemId
        in: path
        description: itemId path parameter.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/projects/{projectId}/quota-lines:
    get:
      operationId: getProjectsByProjectIdQuotaLines
      summary: GET /v1/projects/:projectId/quota-lines
      tags:
        - Quota Lines
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}/quota-lines/{quotaLineId}:
    put:
      operationId: putProjectsByProjectIdQuotaLinesByQuotaLineId
      summary: PUT /v1/projects/:projectId/quota-lines/:quotaLineId
      tags:
        - Quota Lines
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      - name: quotaLineId
        in: path
        description: quotaLineId path parameter.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteProjectsByProjectIdQuotaLinesByQuotaLineId
      summary: DELETE /v1/projects/:projectId/quota-lines/:quotaLineId
      tags:
        - Quota Lines
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      - name: quotaLineId
        in: path
        description: quotaLineId path parameter.
        required: true
        schema:
          type: string
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}/quota-lines/batch-create:
    post:
      operationId: postProjectsByProjectIdQuotaLinesBatchCreate
      summary: POST /v1/projects/:projectId/quota-lines/batch-create
      tags:
        - Quota Lines
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}/quota-lines/candidates:
    get:
      operationId: getProjectsByProjectIdQuotaLinesCandidates
      summary: GET /v1/projects/:projectId/quota-lines/candidates
      tags:
        - Quota Lines
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}/quota-lines/source-chain:
    get:
      operationId: getProjectsByProjectIdQuotaLinesSourceChain
      summary: GET /v1/projects/:projectId/quota-lines/source-chain
      tags:
        - Quota Lines
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}/quota-lines/validate:
    post:
      operationId: postProjectsByProjectIdQuotaLinesValidate
      summary: POST /v1/projects/:projectId/quota-lines/validate
      tags:
        - Quota Lines
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...